Clean up AudioButton class
